Remove hwtracer_private.h.
This header was only used by `collect.c`, so this commit inlines the
header.
