--- hwtracer/src/hwtracer_private.h
@@ -1,24 +1,0 @@
-#ifndef __HWT_PRIVATE_H
-#define __HWT_PRIVATE_H
-
-#include <inttypes.h>
-#include <stdbool.h>
-#include <stddef.h>
-
-enum hwt_cerror_kind {
-  hwt_cerror_unused,
-  hwt_cerror_unknown,
-  hwt_cerror_errno,
-  hwt_cerror_pt,
-};
-
-#define PT_ERROR_OVERFLOW 0
-
-struct hwt_cerror {
-  enum hwt_cerror_kind kind; // What sort of error is this?
-  int code;                  // The error code itself.
-};
-
-void hwt_set_cerr(struct hwt_cerror *, int, int);
-
-#endif
